Added start screen

+++ Raindrops_game/Raindrops_game.pde
@@ -4,6 +4,7 @@ int threshold = 1000;
 Raindrop[] r = new Raindrop[500];
 Catcher catcher;
 int score;
+boolean start = false;
 
 void setup() {
   size(500, 500);
@@ -14,26 +15,37 @@ void setup() {
 }
 
 void draw() {
-  background(200, 230, 255);
-  textSize(50);
-  fill(10);
-  text(score, 10, 100);
-  for (int i = 0; i < index; i++) {
-    r[i].display();
-    r[i].drop();
-    if (catcher.catchDrop(r[i]) == true) {
-      r[i].goAway();
-      score++;
-      threshold-=10;
+  if (!start) {
+    background(0);
+    fill(random(255));
+    ellipse(width/2, height/2, 100, 100);
+  }
+  else {
+    background(200, 230, 255);
+    textSize(50);
+    fill(10);
+    text(score, 10, 100);
+    for (int i = 0; i < index; i++) {
+      r[i].display();
+      r[i].drop();
+      if (catcher.catchDrop(r[i]) == true) {
+        r[i].goAway();
+        score++;
+        threshold-=10;
+      }
+    }
+    catcher.display();
+    catcher.update();
+    if (millis() - oldTime > threshold) {
+      if (index < r.length) {
+        index++;
+        oldTime = millis();
+      }
     }
   }
-  catcher.display();
-  catcher.update();
-  if (millis() - oldTime > threshold) {
-    if (index < r.length) {
-      index++;
-      oldTime = millis();
-    }
-  }
+}
+
+void mousePressed() {
+  start = true;
 }
